moved contents of main.py into __main__.py, deleted main.py
